+++ src/main/java/com/dudus/diecast_api/repository/TransaksiRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TransaksiRepository extends JpaRepository<Transaksi, Long> {
+public interface TransaksiRepository extends JpaRepository<Transaksi, Integer> {
 
     
 }
